// src/interfaces/BackgroundCanvas.interfaces.ts
interface SizeValues {
    width: number
    paddingLeft: number
    paddingRight: number
    borderLeft: number
    borderRight: number
    marginLeft: number
    marginRight: number
    height: number
    paddingTop: number
    paddingBottom: number
    borderTop: number
    borderBottom: number
    marginTop: number
    marginBottom: number
};

export { SizeValues };